Fix fatal error caused by stubs This prevents the following fatal error: ``` PHP Fatal error:  Cannot redeclare sanitize_locale_name() (previously declared in /polylang-wc/vendor/php-stubs/wordpress-stubs/wordpress-stubs.php:104556) in /polylang-wc/vendor/wpsyntex/polylang-stubs/polylang-stubs.php on line 21903 ```
PHP-Stubs doesn't seem to provide a way to exclude functions, only files.
We could move `sanitize_locale_name()` to a dedicated files for WP functions and not include this file in the stubs, but there is simpler way: if a function already exists, it isn't included in the stubs (see `StubsGenerator\NodeVisitor::needsNode()`).
So declaring the function in `finder.php` prevents it from being included in the stubs.